Move signallers to conditions.c
